+++ Frontend/src/routes.jsx
@@ -15,7 +15,7 @@ import CustomTab from './ui/CustomTab'
 import FlashMessage from './ui/messages/flashMessage'
 import PermissionsProvider from './ui/providers/PermissionsProvider'
 import UserProvider from './ui/providers/UserProvider'
-import PageTabs from './ui/Tabs'
+import PageTabs from './ui/PageTabs'
 
 export const BASE_ROUTE = '/competitions/v2'
